move lc and talos
