+++ src/cpg_workflows/query_modules/seqr_loader.py
@@ -8,7 +8,7 @@ import hail as hl
 
 from cpg_utils.config import get_config, reference_path
 from cpg_utils.hail_batch import genome_build
-from src.cpg_workflows.large_cohort import load_vqsr
+from large_cohort import load_vqsr
 from cpg_workflows.utils import checkpoint_hail
 from hail_scripts.computed_fields import variant_id, vep
