+++ src/large_cohort/dataproc_utils.py
@@ -43,7 +43,7 @@ def dataproc_job(
     Submit script as a dataproc job.
     """
     import cpg_workflows
-    from src.cpg_workflows.large_cohort import dataproc_script
+    from large_cohort import dataproc_script
 
     package_path = to_path(cpg_workflows.__file__).parent
     script_path = to_path(dataproc_script.__file__)
